minwoo/minwoo_backup/backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt>=5.3.0
django-cors-headers==4.3.1
PyMySQL==1.1.0
python-dotenv==1.0.0
Pillow>=10.2.0
pandas>=2.1.3
requests>=2.31.0